requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
chromadb>=0.4.0
flask>=3.0.0
gradio>=4.0.0
pypdf>=3.17.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
python-dotenv
ipykernel
duckduckgo-search
langchain-community
langchain-groq
gunicorn>=21.2.0